+++ src/game/levels/level_1.js
@@ -2,8 +2,8 @@ export default {
     name: 'Level 1',
     slug: 'level_1',
     player: {
-        pos_x: 100,
-        pos_y: 100
+        pos_x: 0,
+        pos_y: 0
     },
     units: [],
     objects: [],